optimize last

+++ src/module/post/PostManage.js
@@ -3,6 +3,7 @@ import { Button } from "components/button";
 import { Dropdown } from "components/dropdown";
 import { LabelStatus } from "components/label";
 import { Table } from "components/table";
+import { useAuth } from "contexts/auth-context";
 import { db } from "firebase-app/firebase-config";
 import {
   collection,
@@ -21,9 +22,9 @@ import DashboardHeading from "module/dashboard/DashboardHeading";
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import Swal from "sweetalert2";
-import { postStatus } from "utils/constants";
+import { postStatus, userRole } from "utils/constants";
 
-const POST_PER_PAGE = 5;
+const POST_PER_PAGE = 10;
 
 const PostManage = () => {
   const [postList, setPostList] = useState([]);
@@ -31,6 +32,7 @@ const PostManage = () => {
   const [lastDoc, setLastDoc] = useState();
   const [total, setTotal] = useState(0);
   const navigate = useNavigate();
+  const { userInfo } = useAuth();
   useEffect(() => {
     async function fetchData() {
       const colRef = collection(db, "posts");
@@ -73,9 +75,11 @@ const PostManage = () => {
       cancelButtonColor: "#d33",
       confirmButtonText: "Yes, delete it!",
     }).then(async (result) => {
-      if (result.isConfirmed) {
+      if (result.isConfirmed && userInfo?.role === userRole.ADMIN) {
         await deleteDoc(docRef);
         Swal.fire("Deleted!", "Your post has been deleted.", "success");
+      } else {
+        Swal.fire("Failed!", "You have no right to delete post", "warning");
       }
     });
   }
@@ -117,6 +121,8 @@ const PostManage = () => {
       documentSnapshots.docs[documentSnapshots.docs.length - 1];
     setLastDoc(lastVisible);
   };
+  // const { userInfo } = useAuth();
+  // if (userInfo.role !== userRole.ADMIN) return null;
   return (
     <div>
       <DashboardHeading
